server/marketplace-service/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
requests==2.31.0
psycopg2-binary==2.9.9